+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/PoolPageIT.java
@@ -23,6 +23,7 @@ import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.PoolPage;
 import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.PoolPage.Field;
 import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.dialog.AddNoteDialog;
 import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.element.Note;
+import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.util.TestUtils;
 
 
 public class PoolPageIT extends AbstractIT {
@@ -42,7 +43,7 @@ public class PoolPageIT extends AbstractIT {
     fields.put(Field.NAME, null);
     fields.put(Field.BARCODE, null);
     fields.put(Field.DESCRIPTION, null);
-    fields.put(Field.QC_PASSED, "Unknown");
+    fields.put(Field.QC_STATUS, "Not Ready");
     fields.put(Field.VOLUME, null);
     fields.put(Field.DISCARDED, Boolean.FALSE.toString());
     fields.put(Field.LOCATION, null);
@@ -83,7 +84,7 @@ public class PoolPageIT extends AbstractIT {
     fields.put(Field.PLATFORM, PlatformType.ILLUMINA.getKey());
     fields.put(Field.CONCENTRATION, "6.5");
     fields.put(Field.CREATE_DATE, "2017-08-15");
-    fields.put(Field.QC_PASSED, "False");
+    fields.put(Field.QC_STATUS, "Failed");
     fields.put(Field.VOLUME, "12.0");
     fields.put(Field.DISCARDED, Boolean.FALSE.toString());
     fields.put(Field.LOCATION, null);
@@ -93,7 +94,7 @@ public class PoolPageIT extends AbstractIT {
     changes.put(Field.ALIAS, "1IPO_POOL_1_CHANGED");
     changes.put(Field.DESCRIPTION, "changed desc");
     changes.put(Field.CONCENTRATION, "7.25");
-    changes.put(Field.QC_PASSED, "True");
+    changes.put(Field.QC_STATUS, "Ready");
     changes.put(Field.VOLUME, "0.0");
     page1.setFields(changes);
 
@@ -120,7 +121,7 @@ public class PoolPageIT extends AbstractIT {
     fields.put(Field.PLATFORM, PlatformType.ILLUMINA.getKey());
     fields.put(Field.CONCENTRATION, "6.5");
     fields.put(Field.CREATE_DATE, "2017-08-15");
-    fields.put(Field.QC_PASSED, "Unknown");
+    fields.put(Field.QC_STATUS, "Not Ready");
     fields.put(Field.VOLUME, null);
     fields.put(Field.DISCARDED, Boolean.FALSE.toString());
     fields.put(Field.LOCATION, null);
@@ -129,7 +130,7 @@ public class PoolPageIT extends AbstractIT {
     Map<PoolPage.Field, String> changes = Maps.newLinkedHashMap();
     changes.put(Field.BARCODE, "ITS:A:BAR:CODE");
     changes.put(Field.DESCRIPTION, "added desc");
-    changes.put(Field.QC_PASSED, "True");
+    changes.put(Field.QC_STATUS, "Ready");
     changes.put(Field.VOLUME, "9.99");
     page1.setFields(changes);
 
@@ -156,7 +157,7 @@ public class PoolPageIT extends AbstractIT {
     fields.put(Field.PLATFORM, PlatformType.ILLUMINA.getKey());
     fields.put(Field.CONCENTRATION, "6.5");
     fields.put(Field.CREATE_DATE, "2017-08-15");
-    fields.put(Field.QC_PASSED, "False");
+    fields.put(Field.QC_STATUS, "Failed");
     fields.put(Field.VOLUME, "12.0");
     fields.put(Field.DISCARDED, Boolean.FALSE.toString());
     fields.put(Field.LOCATION, null);
@@ -165,7 +166,7 @@ public class PoolPageIT extends AbstractIT {
     Map<PoolPage.Field, String> changes = Maps.newLinkedHashMap();
     changes.put(Field.BARCODE, null);
     changes.put(Field.DESCRIPTION, null);
-    changes.put(Field.QC_PASSED, "Unknown");
+    changes.put(Field.QC_STATUS, "Not Ready");
     changes.put(Field.VOLUME, null);
     page1.setFields(changes);
 
@@ -318,9 +319,7 @@ public class PoolPageIT extends AbstractIT {
     assertAttribute(Field.PLATFORM, expectedValues, pool.getPlatformType().getKey());
     assertAttribute(Field.CONCENTRATION, expectedValues, LimsUtils.toNiceString(pool.getConcentration()));
     assertAttribute(Field.CREATE_DATE, expectedValues, LimsUtils.formatDate(pool.getCreationDate()));
-    String qcPassed = replaceIfNull(nullOrToString(pool.getQcPassed()), "Unknown");
-    qcPassed = qcPassed.substring(0, 1).toUpperCase() + qcPassed.substring(1);
-    assertAttribute(Field.QC_PASSED, expectedValues, qcPassed);
+    assertAttribute(Field.QC_STATUS, expectedValues, TestUtils.qcPassedToString(pool.getQcPassed()));
     assertAttribute(Field.VOLUME, expectedValues, LimsUtils.toNiceString(pool.getVolume()));
     assertAttribute(Field.DISCARDED, expectedValues, Boolean.toString(pool.isDiscarded()));
     assertAttribute(Field.LOCATION, expectedValues, pool.getLocationBarcode());