+++ miso-web/src/main/webapp/scripts/form_pool.js
@@ -83,7 +83,26 @@ FormTarget.pool = (function($) {
           type: 'date',
           required: true,
           initial: Utils.getCurrentDate()
-        }, FormUtils.makeQcPassedField(true), {
+        }, {
+          title: 'QC Status',
+          data: 'qcPassed',
+          type: 'dropdown',
+          source: [{
+            label: 'Ready',
+            value: true
+          }, {
+            label: 'Failed',
+            value: false
+          }],
+          convertToBoolean: true,
+          getItemLabel: function(item) {
+            return item.label;
+          },
+          getItemValue: function(item) {
+            return item.value;
+          },
+          nullLabel: 'Not Ready'
+        }, {
           title: 'Volume',
           data: 'volume',
           type: 'decimal'