+++ RELEASE_NOTES.md
@@ -2,7 +2,12 @@
 
 Changes:
 
-
+  * Updated terminology for pool QC status to be more consistent with sample, library, and library
+    aliquot QC status:
+    * QC Passed? -> QC Status
+    * True -> Ready
+    * False -> Failed
+    * Unknown -> Not Ready
 
 # 1.14.0